Adjust simple and thread safe tests
